source_discoverer_function/requirements.txt
google-cloud-storage==3.2.0
